Backend/mcp/folders_server/requirements.txt
fpdf
boto3
mcp
python-dotenv
